rootston: break up input.h
