+++ include/rootston/keyboard.h
@@ -1,0 +1,22 @@
+#ifndef _ROOTSTON_KEYBOARD_H
+#define _ROOTSTON_KEYBOARD_H
+
+#include <xkbcommon/xkbcommon.h>
+#include "rootston/input.h"
+
+#define ROOTS_KEYBOARD_PRESSED_KEYSYMS_CAP 32
+
+struct roots_keyboard {
+	struct roots_input *input;
+	struct wlr_input_device *device;
+	struct wl_listener key;
+	struct wl_listener modifiers;
+	struct wl_list link;
+
+	xkb_keysym_t pressed_keysyms[ROOTS_KEYBOARD_PRESSED_KEYSYMS_CAP];
+};
+
+void keyboard_add(struct wlr_input_device *device, struct roots_input *input);
+void keyboard_remove(struct wlr_input_device *device, struct roots_input *input);
+
+#endif
