+++ rootston/cursor.c
@@ -18,6 +18,10 @@
 #include "rootston/input.h"
 #include "rootston/desktop.h"
 #include "rootston/view.h"
+#include "rootston/keyboard.h"
+#include "rootston/pointer.h"
+#include "rootston/tablet_tool.h"
+#include "rootston/touch.h"
 
 const struct roots_input_event *get_input_event(struct roots_input *input,
 		uint32_t serial) {
